refactor(service): move service functions into separate files
Similarly, move test suites into separate files.
This way it's more menagable and easier to navigate.
