+++ test/core/index.js
@@ -1,0 +1,19 @@
+'use strict';
+
+const find = require('./find');
+const get = require('./get');
+const create = require('./create');
+const patch = require('./patch');
+const remove = require('./remove');
+const update = require('./update');
+const raw = require('./raw');
+
+module.exports = {
+  find,
+  get,
+  create,
+  patch,
+  remove,
+  update,
+  raw
+};
